Switched from f64 to f32 floating point numbers; I don't need all the precision. Fixed an issue with the floating point number tracker that keeps track of the X and Y axis. Fixed the collision logic for the move_up sprite method. Fixed an issue with the bullets that were falling short before hitting the top boundry. Added clamping correction logic consequently from the nuances of adding and substracting floats. Added collision logic to the game when a bullet hits an alien, though, not when an alien crashes into a bullet. Added a timer for a bit of a fun factor.

+++ src/entities/shooter.rs
@@ -16,7 +16,7 @@ impl Shooter {
     pub fn new(
         engine: Rc<RefCell<Engine>>,
         position: Vec<Coordinate>,
-        velocity: f64,
+        velocity: f32,
     ) -> Result<Self, Error> {
         Ok(Self {
             sprite: Sprite::new(engine, position, velocity)?,
@@ -27,7 +27,7 @@ impl Shooter {
         let _ = self.sprite.spawn();
     }
 
-    pub fn step(&mut self, key: &str, delta_time: f64) -> Result<State, Error> {
+    pub fn step(&mut self, key: &str, delta_time: f32) -> Result<State, Error> {
         match key {
             "left" => self.sprite.move_left(delta_time),
             "right" => self.sprite.move_right(delta_time),
